requirements.txt
absl-py==2.0.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
colour-demosaicing==0.2.2
colour-science==0.4.1
cycler==0.10.0
efficientnet-pytorch==0.7.1
exceptiongroup==1.1.3
execnet==2.0.2
google-auth==2.23.2
google-auth-oauthlib==1.0.0
grpcio==1.59.0
idna==3.4
imageio==2.9.0
importlib-metadata==6.8.0
iniconfig==2.0.0
kiwisolver==1.3.2
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.4.3
networkx==2.6.3
numpy==1.21.2
oauthlib==3.2.2
packaging==23.2
pandas==1.3.3
Pillow==8.3.2
plotly==5.17.0
pluggy==1.3.0
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==2.4.7
pytest==7.4.3
pytest-xdist==3.3.1
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scipy==1.7.1
six==1.16.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tifffile==2021.8.30
tomli==2.0.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.66.1
transforms3d==0.4.1
typing_extensions==4.8.0
urllib3==2.0.5
Werkzeug==3.0.0
zipp==3.17.0